+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Change Log
 ===========
 0.0.19
 -------
+- Support S&P 500 Historical Annual Returns[`#68 <https://github.com/defeat-beta/defeatbeta-api/issues/68>`_]
+- Support S&P 500 Year CAGR Returns[`#69 <https://github.com/defeat-beta/defeatbeta-api/issues/69>`_]
 
 0.0.18
 -------