+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Change Log
 - Support S&P 500 Year CAGR Returns[`#69 <https://github.com/defeat-beta/defeatbeta-api/issues/69>`_]
 - Support S&P 500 Year CAGR Returns Rolling[`#70 <https://github.com/defeat-beta/defeatbeta-api/issues/70>`_]
 - Support Daily Treasury Yield[`#71 <https://github.com/defeat-beta/defeatbeta-api/issues/71>`_]
+- Support WACC[`#35 <https://github.com/defeat-beta/defeatbeta-api/issues/35>`_]
 
 0.0.18
 -------